main: Add helper function zipArraysInObject
This is similar to python's zip(), except that it takes an
object argument where all values in the object are arrays. It then
turns that single object into an array of objects with the same
keys, but the values are now scalar. If there is a mismatch in
length, we prefer to truncate.