ui: Refactor and clean up the way that UI components are arranged
There was far too much duplication in the model that we were using.

Fundamentally, almost all of our cards follow a layout that
looks similar to this:

DiscoveryFeedActivatableFrame:
  DiscoveryFeedContentCardLayout:
    DiscoveryFeedContentPreview
    DiscoveryFeedAppContentDescription|DiscoveryFeedAppStoreDescription
  DiscoveryFeedVideoCardLayout:
    DiscoveryFeedContentPreview

DiscoveryFeedWordQuotePair

We can share a lot of logic in the former case, especially WRT
frame activation and hover states, as well as image sizing logic.