main: Call through KnowledgeSearch' LoadItem through LaunchViaDBusCall
We're exposing a new D-Bus API on the shell called LaunchViaDBusCall
which essentially forward a D-Bus call on to a given application
and shows a splash screen for it if the application is not open
already.

We're probably going to use something similar in a few places,
so created a function called createMetaCallProxy which does
something similar to Gjs' internal makeProxyWrapper function,
except that it allows the caller to override the mechanism by
which a proxy method is called. createShellAppLauncherMetaCallProxy
is a sort of higher-order specialisation of this that calls a
proxy function 'through' LaunchViaDBusCall on org.gnome.Shell.AppLauncher.

https://phabricator.endlessm.com/T16985