main: Rework query stage to exclusively use promises
We now pervasively use promises everywhere in the query stage, so
that we can composably assemble a final list of models, asynchronously.
This avoids the need for awkward countdowns or callbacks - we just
return a list of promises and each promise returns a pair of either
an error or a list of results which we can traverse. When they're
all resolved, we can simply add the results to the model.