ordering: Add a new model ordering algorithm
Essentially what this algorithm does is alternate between two
patterns for choosing which cards to pick and picking cards from
sequential apps over time. As we get further down the feed we
pick more and more cards from the same app. Then we append all
the remaining installable apps.

This isn't guaranteed to work right now, but it has tests that
confirm that the behaviour is at least correct from a testing
perspective.