app store cards: set width request for description part, less padding
We set the width request to 180px, same width as the thumbnail
of the card. We minimize the padding of the content since
we reduced the overall width. Otherwise we end up overflowing
earlier because of the title width. We do word wrap the title
but of one word in the title is very long we might end up pushing
out the overall width > 180px. Another optimization would be to
reduce the font size of the title.