app-store cards: wrap and ellipsize correctly
Doing wrapping and ellipsizing correct at the same time
needs some logic in place that also take character width
and number of lines into account. I had a look at the
settings from gs-image-tile.ui to find a way of doing
this.

The app label does only ellipsize now and the description
of the app does ellipsize and wrap.