main: Just make discoveryFeedCardsFromQueries concerned with the cards
And not the recommended text or event the models. All it needs to
do is query them and sort them into the right order. From there,
the model and the recommended text are actually properties of the
window, so the window can use the result of that function to
work out what to do.

https://phabricator.endlessm.com/T18975