main: Explicitly correct to 'realize', then use GdkWindow
Previously we were calling gtk_window_get_window during window
construction and that happened to work because of the fact that
we realized a child widget earlier. We shouldn't depend on this
behaviour - the idiomatic way to do it is to connect to
the 'realize' signal, then do the work.

https://phabricator.endlessm.com/T18540