main: Use focus-out-event to detect focus changes, not Wnck
Wnck relies on the window manager setting properties on the root
window and as such is very racy. We can do a better job by just
checking the state of the keyboard focus. Admittedly, this is
the 'old way' of doing things, but it seems to be far more reliable.

https://phabricator.endlessm.com/T18540