feed: Port the card querying logic to C
Add some new types to marshal data about the proxies
and returned cards back and forth between libfeed and the
caller.

FeedKnowledgeAppProxy represents a GDBus proxy connected to some
interface that can provide models for the feed, as well as some
additional information, such as what the source of the proxy is
(eg, the app ID or source name), the type of cards that the
proxy produces and the knowledge search interface it exposes.

FeedOrderableModel wraps a EosDiscoveryFeedBaseCardStore and provides
some additional properties such such as the source of the card
and its type. These properties are only used by the ordering algorithm,
at which point the underlying models are unwrapped and returned.

Use eos_discovery_feed_unordered_results_from_queries to asynchronously
query all the relevant knowledge app proxies and return the result
to the provided callback.

https://phabricator.endlessm.com/T21130