main: Hide the window first, then load the items from the feed
We need to ensure that the focus reverts back to the desktop
first before we start loading items from the feed. Otherwise
we will send the d-bus message to load the item, hide the window
and the focus will change halfway through, causing the speedwagon
window to close before the item has finished loading.

https://phabricator.endlessm.com/T22448