tests: Relax card ordering check
Following tests started failing due to change in the way content
is pushed in GHashTable and later looked up. g_hash_table_iter_next
used to return the object in the manner it was inserted in the table
before but now is no longer true (also, the order wasn't guaranteed).

Failing tests:
1) "eventually take many news cards from the same app"
2) "doesnt repeat content app cards at all"

It looks like we can relax the ordering for these tests as it doesn't
affect the ordering of the card types. It mainly addresses to avoid
redundancy between the same type of cards as far as I can see. Hence,
relax the ordering check for these tests and make it pass.

https://phabricator.endlessm.com/T26331